Move cli commands into jsi::* modules
