+++ crates/ubrn_cli/src/jsi/crossplatform/mod.rs
@@ -3,11 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/
  */
-pub(crate) mod android;
-mod codegen;
-pub(crate) mod crossplatform;
-mod generate;
-pub(crate) mod ios;
+pub(crate) mod codegen;
+pub(crate) mod config;
 
-pub(crate) use codegen::get_files;
-pub(crate) use generate::{CmdArg, TurboModuleArgs};
+pub(crate) use self::codegen::get_files;
+pub(crate) use self::config::TurboModulesConfig;
